Allow usage of groupname in order to reuse class (#68)
It is currently not possible to schedule two different instances of the
same class to run as a scheduled job. That limitation restricts us from
reusing a class since it can only be scheduled once.

This PR enables the possibility to schedule two or more instances of the
same class as long as they have different _group names_ set. It's
backwards compatible and will the current functionality is unchanged
unless a groupname is explicitly set.

**Example usage**
```java
public class GroupNameJob extends Job {

  public StartupJob(String groupName) {
    super(groupName);
  }

  @Override
  public void doJob(JobExecutionContext context) throws JobExecutionException {
  }
}
```

[Link to Quartz tutorial](http://www.quartz-scheduler.org/documentation/quartz-2.x/tutorials/tutorial-lesson-02.html)