Jobfactory take groupName into account
We recently added ability to specify a groupName for a scheduled job.
This means we can schedule multiple instances of the same class on the scheduler, as long as the groupName is different.

When Quartz is requesting an instance of the job to run, the current solution finds the job by class only.
This is not good enough, it also needs to take the groupName into account when comparing the jobs in the bundle.

So this is a fix, in order for the jobfactory to work properly.